Move InternalVmType from modMcf to modBase
