+++ modAionImpl/src/org/aion/zero/impl/db/AionRepositoryCache.java
@@ -15,7 +15,7 @@ import org.aion.log.AionLoggerFactory;
 import org.aion.log.LogEnum;
 import org.aion.base.AccountState;
 import org.aion.mcf.db.ContractDetails;
-import org.aion.mcf.db.InternalVmType;
+import org.aion.base.InternalVmType;
 import org.aion.mcf.db.Repository;
 import org.aion.mcf.db.RepositoryCache;
 import org.aion.mcf.db.TransformedCodeInfo;
