+++ modAionImpl/src/org/aion/zero/impl/db/DetailsDataStore.java
@@ -9,7 +9,7 @@ import org.aion.db.impl.ByteArrayKeyValueDatabase;
 import org.aion.db.impl.ByteArrayKeyValueStore;
 import org.aion.db.store.JournalPruneDataSource;
 import org.aion.db.store.XorDataSource;
-import org.aion.mcf.db.InternalVmType;
+import org.aion.base.InternalVmType;
 import org.aion.precompiled.ContractInfo;
 import org.aion.rlp.RLP;
 import org.aion.rlp.RLPElement;
