+++ modAionImpl/src/org/aion/zero/impl/vm/common/BulkExecutor.java
@@ -10,7 +10,7 @@ import org.aion.base.AccountState;
 import org.aion.base.AionTransaction;
 import org.aion.base.TransactionTypeRule;
 import org.aion.zero.impl.config.CfgFork;
-import org.aion.mcf.db.InternalVmType;
+import org.aion.base.InternalVmType;
 import org.aion.mcf.db.RepositoryCache;
 import org.aion.precompiled.ContractInfo;
 import org.aion.types.AionAddress;
