+++ modAionImpl/test/org/aion/zero/impl/blockchain/TransactionCreateSpecificationTests.java
@@ -21,7 +21,7 @@ import org.aion.fastvm.FastVmResultCode;
 import org.aion.log.AionLoggerFactory;
 import org.aion.log.LogEnum;
 import org.aion.mcf.blockchain.Block;
-import org.aion.mcf.db.InternalVmType;
+import org.aion.base.InternalVmType;
 import org.aion.mcf.db.RepositoryCache;
 import org.aion.types.AionAddress;
 import org.aion.types.InternalTransaction;
