+++ modAionImpl/test/org/aion/zero/impl/db/AionRepositoryCacheTest.java
@@ -15,7 +15,7 @@ import org.aion.base.AccountState;
 import org.aion.db.impl.DBVendor;
 import org.aion.db.impl.DatabaseFactory;
 import org.aion.mcf.db.ContractDetails;
-import org.aion.mcf.db.InternalVmType;
+import org.aion.base.InternalVmType;
 import org.aion.mcf.db.RepositoryCache;
 import org.aion.zero.impl.config.CfgPrune;
 import org.aion.zero.impl.config.PruneConfig;
