+++ modAvmVersion1/src/org/aion/avm/version1/internal/ExternalStateForAvm.java
@@ -6,7 +6,7 @@ import org.aion.avm.stub.IAvmExternalState;
 import org.aion.avm.stub.IEnergyRules;
 import org.aion.avm.stub.IEnergyRules.TransactionType;
 import org.aion.base.AccountState;
-import org.aion.mcf.db.InternalVmType;
+import org.aion.base.InternalVmType;
 import org.aion.mcf.db.RepositoryCache;
 import org.aion.types.AionAddress;
 import org.aion.util.types.ByteArrayWrapper;
