+++ modAvmVersion2/src/org/aion/avm/version2/internal/ExternalStateForAvm.java
@@ -6,7 +6,7 @@ import org.aion.avm.stub.IAvmExternalState;
 import org.aion.avm.stub.IEnergyRules;
 import org.aion.avm.stub.IEnergyRules.TransactionType;
 import org.aion.base.AccountState;
-import org.aion.mcf.db.InternalVmType;
+import org.aion.base.InternalVmType;
 import org.aion.mcf.db.RepositoryCache;
 import org.aion.precompiled.ContractInfo;
 import org.aion.types.AionAddress;
