+++ modMcf/src/org/aion/mcf/db/Repository.java
@@ -4,6 +4,7 @@ import java.math.BigInteger;
 import java.util.Collection;
 import java.util.Map;
 import org.aion.base.AccountState;
+import org.aion.base.InternalVmType;
 import org.aion.types.AionAddress;
 import org.aion.util.types.ByteArrayWrapper;
